Using only one traverse method
Here we have still an issue because the method in the traverse is not doing anything
in case of a leaf and is looping on the names for the children but the logic of the
print is inside the MenuItem* and not external, we need to extract it.
Moreover we are plenty of MenuItems class doing the same thing, we need to have only
one.