Added the second item to the menu list
The tests are passing but the design is rotten in the meanwhile. There is a
repetition on the traversal of the tree that is present on the writers.
There is an unbalance on the treatment of the root folder and of the leafs
of the tree that led to adding a lot of ifs.